+++ build.js
@@ -5,7 +5,6 @@ const del = require('del')
 const tmp = require('tmp-promise')
 const clone = require('git-clone/promise')
 const axios = require('axios')
-const tsFileStruct = require('ts-file-parser')
 
 const aliases = {
   bsc: ['bsc_mainnet'],
@@ -18,19 +17,17 @@ async function build() {
   await del(getBuildLocation('**/*.json'))
 
   const blockchainIdToTokens = {}
-  const blockchainNameToChainId = {}
-  const chainIdToBlockchainName = {}
-  const dexNameToChainIdToRouterAddress = {
+  const dexIdToInfo = {
+    sushiswap: {
+      name: 'SushiSwap',
+      url: 'https://sushi.com',
+      contractsRepositoryUrl: 'https://github.com/sushiswap/sushiswap',
+      contractsRepositoryGitUrl: 'git@github.com:sushiswap/sushiswap.git',
+    },
+  }
+  const dexIdToChainIdToContractNameToContractInfo = {
     sushiswap: {},
     fatex: {},
-  }
-  const dexNameToChainIds = {
-    sushiswap: [],
-    fatex: [],
-  }
-  const dexNameToRouterName = {
-    sushiswap: 'UniswapV2Router02.json',
-    fatex: 'UniswapV2Router02.json',
   }
 
   /* ---
@@ -41,13 +38,60 @@ async function build() {
 
   const allChains = (await axios.get('https://chainid.network/chains.json')).data
 
+  const chainNameToChainId = {}
+
+  allChains.forEach(({ chainId, network }) => {
+    chainNameToChainId[network.toLowerCase()] = chainId
+  })
+
+  /* ---
+    sushiswap/sushiswap
+  --- */
+
+  console.log('Parsing sushiswap/sushiswap')
+
+  let tmpDir = await tmp.dir({ unsafeCleanup: true })
+
+  await clone(dexIdToInfo.sushiswap.contractsRepositoryGitUrl, tmpDir.path)
+
+  const smartContractsDeploymentsLocation = path.join(tmpDir.path, 'deployments')
+
+  fs.readdirSync(smartContractsDeploymentsLocation)
+  .forEach(folder => {
+    if (fs.lstatSync(path.join(smartContractsDeploymentsLocation, folder)).isDirectory()) {
+      const chainId = fs.readFileSync(path.join(smartContractsDeploymentsLocation, folder, '.chainId'), 'utf8')
+
+      fs.readdirSync(path.join(smartContractsDeploymentsLocation, folder))
+      .filter(file => file.endsWith('.json'))
+      .forEach(file => {
+        let json = {}
+
+        try {
+          json = JSON.parse(fs.readFileSync(path.join(smartContractsDeploymentsLocation, folder, file), 'utf8'))
+        }
+        catch (e) {
+          console.log('Error parsing sushiswap/suhsiswap JSON', file)
+        }
+
+        if (!dexIdToChainIdToContractNameToContractInfo.sushiswap[chainId]) {
+          dexIdToChainIdToContractNameToContractInfo.sushiswap[chainId] = {}
+        }
+
+        dexIdToChainIdToContractNameToContractInfo.sushiswap[chainId][file.replace('.json', '')] = {
+          address: json.address,
+          abi: json.abi,
+        }
+      })
+    }
+  })
+
   /* ---
     sushiswap/default-token-list
   --- */
 
   console.log('Parsing sushiswap/default-token-list')
 
-  let tmpDir = await tmp.dir({ unsafeCleanup: true })
+  tmpDir = await tmp.dir({ unsafeCleanup: true })
 
   await clone('git@github.com:sushiswap/default-token-list.git', tmpDir.path)
 
@@ -62,10 +106,7 @@ async function build() {
       console.log('Error parsing sushiswap/default-token-list JSON', file)
     }
 
-    if (json.length) {
-      blockchainIdToTokens[json[0].chainId] = json
-    }
-
+    if (json.length) blockchainIdToTokens[json[0].chainId] = json
   })
 
   await tmpDir.cleanup()
@@ -74,85 +115,73 @@ async function build() {
     sushiswap/sushiswap-sdk
   --- */
 
-  console.log('Parsing sushiswap/sushiswap-sdk')
+  // console.log('Parsing sushiswap/sushiswap-sdk')
 
-  tmpDir = await tmp.dir({ unsafeCleanup: true })
+  // tmpDir = await tmp.dir({ unsafeCleanup: true })
 
-  await clone('git@github.com:sushiswap/sushiswap-sdk.git', tmpDir.path)
+  // await clone('git@github.com:sushiswap/sushiswap-sdk.git', tmpDir.path)
 
-  let chainIdEnumLocation = path.join(tmpDir.path, 'src/enums/ChainId.ts')
-  let chainIdEnum = fs.readFileSync(chainIdEnumLocation).toString()
+  // // const sushsiswapConstants = require(path.join(tmpDir.path, 'src/constants/addresses.ts'))
 
-  tsFileStruct.parseStruct(chainIdEnum, {}, chainIdEnumLocation).enumDeclarations[0].members
-  .forEach(({ name, value }) => {
-    const lowerName = correctBlockchainName(name)
+  // // console.log('sushsiswapConstants', Object.keys(sushsiswapConstants))
 
-    blockchainNameToChainId[lowerName] = value
-    chainIdToBlockchainName[value] = lowerName
-    dexNameToChainIds.sushiswap.push(value)
-  })
+  // // dexNamIdhainIdToContractTypeToContract.sushiswap
 
-  dexNameToChainIdToRouterAddress.sushiswap = require(path.join(tmpDir.path, 'src/constants/addresses.ts')).ROUTER_ADDRESS
-
-  await tmpDir.cleanup()
+  // await tmpDir.cleanup()
 
   /* ---
     FATEx-DAO/fatex-dex-sdk
   --- */
 
-  console.log('Parsing FATEx-DAO/fatex-dex-sdk')
+  // console.log('Parsing FATEx-DAO/fatex-dex-sdk')
 
-  tmpDir = await tmp.dir({ unsafeCleanup: true })
+  // tmpDir = await tmp.dir({ unsafeCleanup: true })
 
-  await clone('git@github.com:FATEx-DAO/fatex-dex-sdk.git', tmpDir.path)
+  // await clone('git@github.com:FATEx-DAO/fatex-dex-sdk.git', tmpDir.path)
 
-  chainIdEnumLocation = path.join(tmpDir.path, 'src/constants.ts')
-  chainIdEnum = fs.readFileSync(chainIdEnumLocation).toString()
+  // const chainIdEnumLocation = path.join(tmpDir.path, 'src/constants.ts')
+  // const chainIdEnum = fs.readFileSync(chainIdEnumLocation).toString()
 
-  tsFileStruct.parseStruct(chainIdEnum, {}, chainIdEnumLocation)
-  .enumDeclarations
-  .find(({ name }) => name === 'ChainId')
-  .members
-  .forEach(({ name, value }) => {
-    const lowerName = correctBlockchainName(name)
+  // tsFileStruct.parseStruct(chainIdEnum, {}, chainIdEnumLocation)
+  // .enumDeclarations
+  // .find(({ name }) => name === 'ChainId')
+  // .members
+  // .forEach(({ name, value }) => {
+  //   const lowerName = correctBlockchainName(name)
 
-    if (!blockchainNameToChainId[lowerName]) blockchainNameToChainId[lowerName] = value
-    if (!chainIdToBlockchainName[value]) chainIdToBlockchainName[value] = lowerName
-    dexNameToChainIds.fatex.push(value)
-  })
+  //   if (!blockchainNameToChainId[lowerName]) blockchainNameToChainId[lowerName] = value
+  //   if (!chainIdToBlockchainName[value]) chainIdToBlockchainName[value] = lowerName
+  //   dexNameToChainIds.fatex.push(value)
+  // })
 
   /* ---
     FATEx-DAO/fatex-dex-protocol
   --- */
 
-  console.log('Parsing FATEx-DAO/fatex-dex-protocol')
+  // console.log('Parsing FATEx-DAO/fatex-dex-protocol')
 
-  tmpDir = await tmp.dir({ unsafeCleanup: true })
+  // tmpDir = await tmp.dir({ unsafeCleanup: true })
 
-  await clone('git@github.com:FATEx-DAO/fatex-dex-protocol.git', tmpDir.path)
+  // await clone('git@github.com:FATEx-DAO/fatex-dex-protocol.git', tmpDir.path)
 
-  const smartContractsDeploymentsLocation = path.join(tmpDir.path, 'deployments')
+  // smartContractsDeploymentsLocation = path.join(tmpDir.path, 'deployments')
 
-  fs.readdirSync(smartContractsDeploymentsLocation)
-  .forEach(folder => {
-    if (fs.lstatSync(path.join(smartContractsDeploymentsLocation, folder)).isDirectory()) {
-      const json = require(path.join(smartContractsDeploymentsLocation, folder, 'UniswapV2Router02.json'))
+  // fs.readdirSync(smartContractsDeploymentsLocation)
+  // .forEach(folder => {
+  //   if (fs.lstatSync(path.join(smartContractsDeploymentsLocation, folder)).isDirectory()) {
+  //     const json = require(path.join(smartContractsDeploymentsLocation, folder, 'UniswapV2Router02.json'))
 
-      dexNameToChainIdToRouterAddress.fatex[blockchainNameToChainId[correctBlockchainName(folder)]] = json.address
-    }
-  })
+  //     dexNameToChainIdToRouterAddress.fatex[blockchainNameToChainId[correctBlockchainName(folder)]] = json.address
+  //   }
+  // })
 
   /* ---
     save
   --- */
 
   saveJson(getBuildLocation('allChains.json'), allChains)
-  saveJson(getBuildLocation('blockchainIdToTokens.json'), blockchainIdToTokens)
-  saveJson(getBuildLocation('blockchainNameToChainId.json'), blockchainNameToChainId)
-  saveJson(getBuildLocation('chainIdToBlockchainName.json'), chainIdToBlockchainName)
-  saveJson(getBuildLocation('dexNameToChainIds.json'), dexNameToChainIds)
-  saveJson(getBuildLocation('dexNameToChainIdToRouterAddress.json'), dexNameToChainIdToRouterAddress)
-  saveJson(getBuildLocation('dexNameToRouterName.json'), dexNameToRouterName)
+  saveJson(getBuildLocation('dexIdToInfo.json'), dexIdToInfo)
+  saveJson(getBuildLocation('dexIdToChainIdToContractNameToContractInfo.json'), dexIdToChainIdToContractNameToContractInfo)
 
   Object.entries(blockchainIdToTokens)
   .forEach(([blockchainId, tokens]) => {