+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## v3.0.0
+
+- Removed most files
+- Added `dexIdToChainIdToContractNameToContractInfo.json` and `dexIdToInfo.json`
+
 ## v2.0.0
 
-Moved data to `data` directory.
+- Moved data to `data` directory.
 
 ## v1.1.0
 
-Added `allChains.json`.
+- Added `allChains.json`.